feat: Remove APIs in favour of direct loader usage
- Removed API files: `examples/basic/app/api/feed+api.tsx`, `examples/basic/app/api/notifications+api.tsx`, `examples/basic/app/api/posts+api.ts`, `examples/basic/app/api/profile+api.tsx`
- Updated loader functions in `examples/basic/app/(feed)/index.tsx`, `examples/basic/app/(feed)/post/[id].tsx`, `examples/basic/app/notifications+spa.tsx`, `examples/basic/app/profile.tsx` to use direct database queries
- Deleted data fetching files: `examples/basic/data/feed.tsx`, `examples/basic/data/notifications.tsx`, `examples/basic/data/post.tsx`, `examples/basic/data/profile.tsx`
- Removed `@tanstack/react-query` dependency from `examples/basic/package.json`

Implementation reason: This change simplifies the data fetching mechanism by removing intermediate API layers and directly using loader functions with database queries, improving performance and maintainability.
