+++ examples/basic/package.json
@@ -18,7 +18,6 @@
     "@tamagui/image-next": "^1.110.1",
     "@tamagui/lucide-icons": "^1.110.1",
     "@tamagui/vite-plugin": "^1.110.1",
-    "@tanstack/react-query": "^5.53.3",
     "@vxrn/color-scheme": "1.1.203",
     "better-sqlite3": "^11.2.1",
     "dotenv": "^16.4.5",
